#40 포인트, 수익 패키지 재구성 (#43)
* feat: point -> order, transaction 패키지 변경

* feat: revenue -> settlement 패키지 변경
