+++ src/main/java/com/pentoryall/order/service/OrderService.java
@@ -1,4 +1,4 @@
-package com.pentoryall.point.service;
+package com.pentoryall.order.service;
 
 import org.springframework.stereotype.Service;
