+++ src/main/java/com/pentoryall/settlement/controller/SettlementController.java
@@ -1,4 +1,4 @@
-package com.pentoryall.point.controller;
+package com.pentoryall.settlement.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
@@ -7,5 +7,5 @@ import org.springframework.web.bind.annotation.RequestMapping;
 @Controller
 @RequestMapping
 @RequiredArgsConstructor
-public class OrderController {
+public class SettlementController {
 }
