+++ src/main/java/com/pentoryall/transaction/controller/TransactionController.java
@@ -1,4 +1,4 @@
-package com.pentoryall.revenue.controller;
+package com.pentoryall.transaction.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
@@ -7,5 +7,5 @@ import org.springframework.web.bind.annotation.RequestMapping;
 @Controller
 @RequestMapping
 @RequiredArgsConstructor
-public class SettlementController {
+public class TransactionController {
 }
